backend/requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-filter==23.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
graphene==3.2.2
graphene-django==3.0.1
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
openapi-codec==1.3.2
promise==2.3
psycopg2-binary==2.9.6
pytz==2023.3
PyYAML==6.0
requests==2.30.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
uritemplate==4.1.1
urllib3==2.0.2
